% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw09-package.r
\docType{package}
\name{hw09}
\alias{hw09}
\alias{hw09-package}
\title{hw09: a mini R package by antonyagonzalez}
\description{
The following package is an excerpt of a larger package I am working on focusing on
making regression coefficients easier to deal with. Looking over regression output
and trying to extract what you need from it can be quite the headache, especially if
you have many different models to deal with. This mini package will extract betas
and their relevant statistics from 3 way interaction regression models to get
information for a 2 way interaction between your 1st and 2nd predictor (interact12),
information for the 3 way interaction (interact3), and will meta-analyze standardized
regression coefficients that you input.
}

