% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact3.R
\name{interact3}
\alias{interact3}
\title{Extract regression model data for 3 way interaction}
\usage{
interact3(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value, ci95_low, ci95_high
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the three way interaction specified in your linear model. Linear model must have 3+ predictors to work.
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
interact3(example)
}

