% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabeta.R
\name{metabeta}
\alias{metabeta}
\title{Meta-analyze two standardized regression coefficients (input)}
\usage{
metabeta(b1, se1, b2, se2)
}
\arguments{
\item{b1}{coefficient 1}

\item{se1}{standard error of coef 1}

\item{b2}{coefficient 2}

\item{se2}{standard error of coef 2}
}
\value{
beta, se, z-score, p-value, ci95
}
\description{
If your regression coefficients are standardized, input the
coefficients and standard errors for each to compute a meta-analyzed beta
based off of Lipsey & Wilson (2001). P-values based off of 2-sided test.
Designed for meta analyzing two regression coefficients only.
}
\examples{
example
}

