% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact12.R
\name{interact12}
\alias{interact12}
\title{Extract 3-way interaction regression model data
(interaction between predictor 1 & 2)}
\usage{
interact12(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value, ci95_low, ci95_high
}
\description{
Pull the regression coefficient, standard error, t-value, p-value and confidence
interval for the interaction between the first and second predictors
specified in your linear model. Linear model must have 3+ predictors to work.
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
interact12(example)
}

